[DELETE] imageService 추상화 삭제 #348"
